requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
python-pptx
pydantic
Jinja2
groq
